FLEDGE: Adjust currency for reportResult to bidder currency
We got feedback that it would be preferred, and none of the reason for things to be in seller currency apply (and the bidder can do conversion if it wants to).

The accompanied tag is again based on config, for similar redaction reasons, like in reportWin.

(This is tracked in https://github.com/WICG/turtledove/issues/166)

Change-Id: I733cdcd0ec47dd7166f08ea82e43d8aea2025da2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4573913
Reviewed-by: Russ Hamilton <behamilton@google.com>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151414}