[rust] Enable chrome://crash/rust
This enables some bytes of Rust in the real Chromium executable for the
first time. Those bytes are a handler for chrome://crash/rust URIs, and
thus no Rust will be executed by normal Chrome users. This gives us a
low-risk opportunity to discover any implications of Rust within Chrome
and gives us a chance to test out Rust's crash handling.

For now, we're enabling this only on Android and Linux - this condition
will be relaxed in future.

Bug: 1292038, 1368726
Change-Id: Ia648ee4b3c104d67bed3d7444325453bc7b865ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4386637
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151428}