OOP-VD: Delay the reset() of the StableVideoDecoderFactory.
CL:4538864 made the browser process reset() the
StableVideoDecoderFactory connection once all the StableVideoDecoders
created from it disconnect. When the StableVideoDecoderFactory is
reset(), the remote video decoder process dies. This may cause some
undesired churn if the client creates a StableVideoDecoder, destroys it,
and then creates another StableVideoDecoder. Depending on the timing,
this would end up creating two video decoder processes: the first one to
host the first StableVideoDecoder and the second one to host the second
StableVideoDecoder (and the first process is destroyed). We should be
re-using the same process since the renderer hasn't changed.

This CL accomplishes that by adding a delay before we reset() the
StableVideoDecoderFactory. If all the StableVideoDecoders disconnect but
a request to create another one is received soon after that, we don't
reset() the StableVideoDecoderFactory.

Bug: 1444970, b:195769334
Test: on brya, crosvideo.appspot.com on lacros-chrome while monitoring the task manager and switching between codecs
Test: this CL adds browser tests
Change-Id: Id6c0f7676a34e6d1d9038c93edfe319df46dbe32
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4553359
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151459}