Experiment with caching frequently used KeyedServices
This experiment uses a pointer stored in Profile for a few KeyedServices
that are frequently queried in Chrome. TemplateURLService,
InstantService, and ThemeService are queried thousands of times during
Chrome and having a slightly faster cached pointer may slightly improve
performance in Chrome.

Based on local (linux) builds, getting the KeyedService through this caching mechanism is around 3x faster, and may result in around 1% performance improvement WRT CPU cycles. However, this may be an overstatement based on the configuration of local builds. This may translate to performance on Windows, but it seems unlikely Android relies so heavily on KeyedServices.

Other possible KeyedServices that are queries often are:
extensions::PermissionHelper
extensions::ProcessMap
extensions::ExtensionRegistry

However, these should not be stored in Profile due to layering
violations.

Bug: b/283002660
Change-Id: I91c83cb5e0402741270d5db4528eddfe02cb3d12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4540880
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Ryan Sturm <ryansturm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151463}