Avoid `raw_ptr` hiding in system_dialog_delegate_view.h
The `raw_ptr` variable is hiding the `raw_ptr` type. This is not a
problem in general. Removing it is not useful in general, but it would
make my life easier. Indeed, I am using locally a tweaked version of the
`raw_ptr` class (base::raw_ptr). It becomes a macro and it is used to
automatically list the location of every dangling pointers in Chrome.
Something along the lines of: https://godbolt.org/z/o8adYnc7a

By applying this patch, I won't have to apply this manually anymore.

Bug: None
Change-Id: I61f402e09df939f7a37be3ecdf5f1edc8a699e8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4567486
Commit-Queue: Xiaodan Zhu <zxdan@chromium.org>
Reviewed-by: Xiaodan Zhu <zxdan@chromium.org>
Auto-Submit: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151473}