cros_next: create MaterialNext scroll bar
This cl changes current RoundedScollBar to fit the MaterialNext
design spec when the flag Jellyroll is enabled. The main changes
are listed below:
- Make the scroll bar thinner when it is not hovered or pressed.
- Add an outline to the thumb.
- Change the color ID to dynamic color tokens.

Visually tested on style viewer and launcher. The effects are
shown in the bug comments.

Change the rounded scroll bar unittests to include the Jellyroll
enabled cases.

Test: *RoundedScrollBarTest*
Bug: b:279499593
Change-Id: I2d78f19893667f3c1a306ec12abf02976f8912dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4575444
Commit-Queue: Xiaodan Zhu <zxdan@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151481}