Add additional bubble to password manager tutorial
This adds an element identifier to the OK button on the PWA install bubble dialog. It also emits a custom event when the PWA has been installed.

This allows us to anchor a help bubble to the dialog for the Password Manager tutorial. The help bubble will show anchored to the Ok button and the tutorial will complete when the user has installed the Password Manager PWA.

Bug: 1445187
Change-Id: I231abe9709faca5f30acea37986e39c1538222ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4575643
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Commit-Queue: Mickey Burks <mickeyburks@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151512}