Replace ClientCertType with the signature algorithm list
The client cert type was removed in TLS 1.3. In TLS 1.2, both it and the
signature algorithm list exist. In TLS 1.0 and 1.1, we only have client
cert type.

We no longer support TLS 1.0 and 1.1, so migrate to the new one. This
only impacts Android. On other platforms, we ignore the field anyway.
Android wants key type values as strings, so add a common helper
function to do the conversion. In doing so, this fixes a bug in the
WebView copy, where it used "ECDSA" instead of "EC". This was fixed
in the Chrome copy in https://crrev.com/c/1863377, but I must have
missed that there was a second copy.

Bug: 1270530
Change-Id: Iddc8b0b864bd2b2db54885398f7cf4f2a2352926
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4528838
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Auto-Submit: David Benjamin <davidben@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Matt Mueller <mattm@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151548}