Ensure initialization parity between DirtinessState and ShouldSave
TabState and CriticalPersistedTabData have DirtinessState and
ShouldSave respectively to throttle saves and ensure saves don't
get too chatty. It was observed that in some instances the first
CriticalPersistedTabData save was being missed. Upon investigation,
it was determined to be because during TabStateAttributes
initialization DirtinessState was set to DirtinessState.DIRTY without
the equivalent shouldSave being set to true. This CL ensures
parity in that scenario.

Bug: 1450393
Change-Id: I811a48766f571e70769bb629df3e4c4f94fc4280
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4576161
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: David Maunder <davidjm@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1151575}