[iOS] Remove duplicate code path for reordering overflow menu
This CL removes the duplicate code path for reordering badged overflow
menu items (one for with smart sorting, one for without). It uses the
following principles:

- NSArray<OverflowMenuDestination*> is solely used to communicate above
this class. Within this class and below this class, it uses
DestinationRanking (= std::vector<overflow_menu::Destination>) to hold
destination orderings.
- The _ranking instance variable is modified instead of passing around
vectors

Now, the behavior is as follows:
1. Use provided default ranking as a default if there is no ranking
currently (either because this is the first use of the menu or user
never used smart sorting in the past).
2. Ask Smart Sorting to sort, if available.
3. Reorder the menu due to badging, if necessary (in the future, this
could be refactored into a separate module, like Smart Sorting).

Bug: 1446779
Change-Id: I658523aaed5775b1dde3117ff1622dafb1818a28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4563735
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Benjamin Williams <bwwilliams@google.com>
Cr-Commit-Position: refs/heads/main@{#1151589}