media/gpu/test: Introduce RawVideo
media::test::Video is too messy because it manages the compressed
video data and the raw video data. We are going to decouple the class
to two classes, one for compressed video only and the other one for
the raw video only.
This CL introduces the media::test::Raw. It manages the raw video.

Bug: b:282878791
Test: Build
Change-Id: Ibcd40d44a832ecab5d40fc2be51f6288ae4717da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4547035
Reviewed-by: Justin Green <greenjustin@google.com>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151609}