Ensure the app "New window" submenu has a radio button selected.
In tabbed mode, if you enable the flag kDesktopPWAsTabStripSettings, a
third option gets added to this submenu. But when the flag is off,
previously the hidden extra item was selected, so neither visible
option was shown as selected.

The kTabbedWindow window mode should only be used when the flag is
enabled.

Before and after screenshots: https://crbug.com/1448741#c1

Bug: 1448741
Change-Id: If07e018ead68c04b73972540aff655893a358efa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4575626
Commit-Queue: Louise Brett <loubrett@google.com>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151668}