[iOS][Signin] Replace DCHECK by NOTREACH to a crash report in stable
This DCHECK() failure seems to be impossible to reproduce with canary
and dev. But it seems to happen with stable.
To have a crash with a stack trace, this DCHECK is changed into
a NOTREACHED() which does sent a crash report.

This patch needs to be cherry-picked in M115 and M114 if possible.

Fixed: 1450372
Change-Id: I0ea0f9c13c69ae02f3648b875eff0bbb9ba92fe2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4575210
Auto-Submit: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151757}