Include AllocationStackTraceRecorder into Crashpad report - 3/N
With Arm's Memory Tagging Extension enabled, a process shall receive a
SIGSEGV if it encounters a mismatch in memory tags. To help developers
analyse problems like this, we want to include a history of the most
recent allocations and frees in the crashpad report.

This CL extends the allocation_recorder component. It now locates the
AllocationTraceRecorder from the terminating process and copies it into
the crash handler. The crashpad report still only contains simple
messages.

Bug: 1137393
Change-Id: I3dcc039e24e1ac201d145d1de153cfbec4fb5383
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4442107
Commit-Queue: Andre Kempe <andre.kempe@arm.com>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Richard Townsend <richard.townsend@arm.com>
Cr-Commit-Position: refs/heads/main@{#1151761}