Kcer-over-NSS: Fix queue unblocker and ListCerts
* When the IO thread gets stopped, tasks for it are destroyed on
their current threads. ScopedClosureRunner on destruction checks
whether its closure is still valid and if it is doing that on a
wrong thread, it effectively derefences the weak pointer from
the callback on a wrong thread and crashes on the invalid
sequence check.Make ScopedClosureRunner from KcerTokenImplNss
unconditionally post its closure on the correct thread instead
of trying to run it. In most cases it will work the same, in
the described scenario this will destroy the closure without
checking the weak pointer.
* Check the cert_list pointer before iterating over the list in
the ListCertsOnWorkerThread method.

Bug: b:244408716
Test: Kcer* unit tests
Change-Id: I690722da62739c744531876966e271864e8e6f5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4568007
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Michael Ershov <miersh@google.com>
Cr-Commit-Position: refs/heads/main@{#1151877}