browser: deletes obsolete test
PermissionPromptBubbleBaseViewBrowserTest.SwitchBrowserWindow
was primarily added to exercise some mac code that has since been
deleted. This test is problematic as it ends up adding a WebContents
to a Browser that is about to be deleted. I'm in the process of
making Browser CHECKs in this scenario. As the test is no longer
applicable and doing something we don't want to support I'm going
to remove it.

Bug: 1434387

Change-Id: If3b3d6337947fbc875115049cec4b2d84d332ca3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4578124
Reviewed-by: Trent Apted <tapted@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151884}