[Payments] More verbose developer error message for manifest downloads
Before this patch, whenever Chrome failed to download a payment
manifest, the PaymentRequest.show() JS API returned a generic error
message "Unable to download payment manifest <URL>."

The generic error message did not explain the reason for the failure,
which made it difficult for web developers to debug what happened. For
example, the payment manifest downloader has a 1MB limit, but the
generic error message did not reveal to the web developer when that
limit was being hit.

This patch extends the error message to provide either HTTP or network
error message or code. For example:

    Unable to download payment manifest <URL>.
        ERR_INSUFFICIENT_RESOURCES (-12)

    Unable to download payment manifest <URL>.
        HTTP 404 Not Found.

Change-Id: I0f82bba6dc26f9edbc11ce5632bd77df5527b127
Bug: 1450588
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4566412
Reviewed-by: Nick Burris <nburris@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151914}