Add NV12_MULTIPLANE support to CopyOutputNV12() for blit requests
This CL adds support for NV12_MULTIPLANE to CopyOutputNV12() in the case
where a blit request is present and we are waiting for GPU work to
complete (the only case in which requests with NV12_MULTIPLANE format
are currently generated). This mostly involves parameterizing over the
number of mailboxes that should be interacted with throughout the flow.
The true meat of the change is where we obtain both surfaces from the
single multiplanar SI representation rather than the per-plane
representations.

This CL does not yet add support for the case of creating surfaces or
for the case where CopyOutputNV12() is not waiting for GPU work to
complete, but rather CHECK()s that these cases do not yet use
NV12_MULTIPLANE. Adding support for these and then enabling them via
FrameSinkVideoCapturerImpl is followup work.

Bug: 1429004
Change-Id: Ie6f9e8b7e97ef2e33ed1d69f54c304dfa83b1752
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4565237
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151934}