[dPWA] Observe OnRegistrationStored() for service workers for ML promo
This CL modifies the MLInstallabilityPromoter to observe
OnRegistrationStored() for service workers instead of
OnRegistrationCompleted(). Since the former is always fired after the
latter, the service worker is always guaranteed to be registered.

This also ensures that the data returned by the QuotaProxyManager
works properly, and fixes the flakiness in the SiteQualityMetrics tests.

Bug: 1450421, b/284157768
Change-Id: I237a948d92b5ced37d333d0e58d540e153e3122b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4580548
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151936}