Fixed "Some tabs are showing “x” and can be closed during selection"
The cause of the issue is that when `TabGridMode` is changed the GridCell's `state` property is changed only for the visible cells. However, in some cases `UICollectionView` dequeues and configures more cells than are currently visible on the screen. Thus, those cells are never got updated about the `TabGridMode` change.

The fix updates the cells' `state` property if needed before displaying them.

Fixed: 1427278
Change-Id: I358a6923bbf9559fa51882ad50f9bdbf7c78d9da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4577298
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Petro Akzhygitov <pakzhygitov@google.com>
Cr-Commit-Position: refs/heads/main@{#1151967}