Track parent_clip_rect across frames in SurfaceAggregator.
This is for the feature AllowUndamagedNonrootRenderPassToSkip.

Introduce |parent_clip_rect| which is the intersection of ancestors'
|output_rect|, RenderPassDrawQuad rect, SurfaceDrawQuad rect and |clip
rect|. If |parent_clip_rect| changes from the previous frame, this
render pass cannot be skipped.

If a render pass is merged to its parent render pass in the previous
frame but is not merged in the current frame, it has to redraw with full
damage because there is no separate output buffer for this render pass
when it's merged.

Remove in_pixel_moving_filter_pass from
SurfaceAggregator::RenderPassNeedsFullDamage(). The damage rect of a
render pass with filters is calculated correctly in Prewalk. It doesn't
need to full damage.

This implementation of tracking render pass parent_clip_rect across
frames fixes the regressions in linux WPT tests when a skipped render pass is allow.

Bug: 1446098
Change-Id: I0300c88921fad20612936a3ece3648782c076d5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4565025
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151977}