[iOS] Replace WebState::GetCurrentURL with GetLastCommittedURLIfTrusted
WebState::GetCurrentURL is a confusing API that doesn't match
any concepts in WebContents. Callers typically instead need the last
committed URL.

Since WebKit doesn't directly expose a last committed URL, a notion
of trust is still needed, since there are cases where WebState can't
be certain about the true last committed URL.

This CL introduces a GetLastCommittedURLIfTrusted method on WebState
and migrates most callers of GetCurrentURL to this new method. This
method returns an absl::optional<GURL> which is absl::nullopt when
the last committed URL is not trusted.

NavigationManagerImpl::GetLastCommittedItemInCurrentOrRestoredSession
continues to need GetCurrentURL but doesn't use the `trust_level`
argument passed to it, so this CL moves GetCurrentURL to
NavigationManagerDelegate where it can only be used by
NavigationManagerImpl.

With these changes, the URLVerificationTrustLevel concept becomes
unused, so this CL deletes this, leading to more simplification in
CRWWebController.

Bug: 457679
Change-Id: If4d584e47490232ae72025005146e0268ddb6d15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4571221
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151988}