Fix SessionScopeSettingsRestoreSession test
This test was passing in a wrong way by checking the wrong object. It
cannot be fixed by just checking the correct object. Getting rid of the
`PrefProvider` created when creating the TestingProfile is the key.

This is because the test creates multiple PrefProvider instances (one
implicitly created by TestingProfile, others created in the test to
simulate shutdown and restart of Chrome), thus there are multiple
ContentSettingsPref instances created for the same
ContentSettingsType. When one of the ContentSettingsPref instances
makes a pref change, the other instance will be propagated with the
change, and because the other instance does not have
`updating_preferences_ == true`, it would proceed to
`ReadContentSettingsFromPref();` and reload prefs. For this test case,
the `PrefProvider` created when creating the TestingProfile can
delete the settings change made by `provider` from prefs (since it's
non-durable), so that `provider2` will not be able to read that
settings value.

Bug: 1399301
Change-Id: I74f1d6bcf3d5420661ec5d33d84392a32dac96ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4089408
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Shuran Huang <shuuran@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151991}